pyproject.toml
[project]
name = "youtube-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "faiss-cpu>=1.7.3",
    "groq>=0.3.1",
    "huggingface-hub>=0.14.1",
    "langchain>=0.0.200",
    "langchain-community>=0.0.10",
    "numpy>=1.24.0,<2.0.0",
    "python-dotenv>=1.0.0",
    "pytube==15.0.0",
    "requests>=2.28.2",
    "sentence-transformers>=2.2.2",
    "streamlit>=1.24.0",
    "streamlit-chat==0.1.0",
    "streamlit-extras==0.3.1",
    "transformers>=4.30.0",
    "youtube-transcript-api>=0.6.1",
]
